refactor(pipeline): extract brief templates into briefs.py (SRP)
implementer.py mixed two responsibilities: the execution strategies
(Claude/Worker implementers + factory) and ~340 lines of imperative
PT-BR prompt templates with their renderers. Move all _WORKER_*_BRIEF
constants, the mention-rendering helpers and _render_claude_mention_prompt
into a dedicated briefs.py, mirroring how claude_dispatcher separates its
prompt templates from dispatch logic. implementer.py now owns only the
strategies and imports the renderers from briefs.
